travisrequirements.txt
South==0.8.4
django-classy-tags==0.5.1
django-cms==2.4.3
django-mptt==0.5.2
django-nose==1.2
django-sekizai==0.7
html5lib==0.999
nose==1.3.3
six==1.6.1
wsgiref==0.1.2
